+++ R/utils-expr.R
@@ -15,7 +15,7 @@ node_walk_replace <- function(node, old, new) {
   while(!is_null(node)) {
     switch_expr(node_car(node),
       language = node_walk_replace(node_cdar(node), old, new),
-      symbol = if (identical(node_car(node), old)) set_node_car(node, new)
+      symbol = if (identical(node_car(node), old)) mut_node_car(node, new)
     )
     node <- node_cdr(node)
   }
@@ -23,7 +23,7 @@ node_walk_replace <- function(node, old, new) {
 expr_substitute <- function(expr, old, new) {
   expr <- duplicate(expr)
   switch_type(expr,
-    quote = ,
+    quosure = ,
     language = node_walk_replace(node_cdr(expr), old, new),
     symbol = if (identical(expr, old)) return(new)
   )