+++ inst/include/dplyr/Result/GroupedSubset.h
@@ -86,7 +86,10 @@ namespace dplyr {
     typedef typename Rcpp::traits::storage_type<RTYPE>::type STORAGE;
 
     SummarisedSubsetTemplate(SummarisedVariable x) :
-      object(x), output(1) {}
+      object(x), output(1)
+    {
+      copy_most_attributes(output, object);
+    }
 
     virtual SEXP get(const SlicingIndex& indices) {
       output[0] = object[indices.group()];
@@ -105,24 +108,9 @@ namespace dplyr {
   };
 
   template <>
-  class SummarisedSubsetTemplate<VECSXP> : public GroupedSubset {
-  public:
-    SummarisedSubsetTemplate(SummarisedVariable x) : object(x) {}
-
-    virtual SEXP get(const SlicingIndex& indices) {
-      return List::create(object[indices.group()]);
-    }
-
-    virtual SEXP get_variable() const {
-      return object;
-    }
-    virtual bool is_summary() const {
-      return true;
-    }
-
-  private:
-    List object;
-  };
+  inline SEXP SummarisedSubsetTemplate<VECSXP>::get(const SlicingIndex& indices) {
+    return List::create(object[indices.group()]);
+  }
 
   inline GroupedSubset* summarised_subset(SummarisedVariable x) {
     switch (TYPEOF(x)) {