+++ NEWS.md
@@ -8,6 +8,10 @@
   The default can also be tweaked by calling
   `pkgconfig::set_config("dplyr::na_matches", "na")` (#2033).
 
+* `summarise()` now correctly evaluates newly created factors (#2217).
+
+* `summarise()` now can create ordered factors (#2200).
+
 * `bind_rows()` and `bind_cols()` give an error for database tables (#2373).
 
 * `bind_rows()` works correctly with `NULL` arguments and an `.id` argument